Patch traindate,score calculation, absolute path, how to use
Until now I was working with absolute path, this was corrected to relative path, be aware of your working directory while running the file.
The calculation of the score was wrong, I was scanning the list in the wrong direction.
Train function was defining the rows before the loop however a word next_revision can be set to "now" so if n was superior to the number of word with a past next revision we were looking to later one while the could be the same word again.
I added a partial explanation to how to use the app right now (the command in the promt are not completed) and file format is still not described