Db restructuration
In order to allow translation to any language to another without creating dozen of table (something necessary on the previous db structure) I restructured all the db.
Now there is one table per languages
(currently only languages supported by wordreference and a na table for other languages)
A table trad which is build more efficiently regrouping previous trad and train table with an additionnal attribute, groups which will be usefull while creating set of word by theme.
The same table stats.

I also corrected the problem of a word with only 1 translation which was just skipped in post function.

There was an other issue, I forgot a return in training which lead to to much words delivered