Automatic translation for every languages
Now automatic translation work for any language suported by wordreference (be aware than russian to english is not supported by wordreference, they use other dictionnaries only so it do not work).
Iceland issue was corrected,
Chinese issue was corrected,
Many translation with the same meaning was partly corrected, now meaning separated by a coma is supported but sometime there different translation on multiple line for different reason, currently only the first line is taken into account.
Be aware that wordreference show translation on from language a to language b but also to language b to language a on the same page (it's the case mainly on language with smaller database), this is not supported for the moment.
I had forgotten arabic in previous commit it's added now.
Automatic translation is still not implemented to work while adding words to the database, i make it clean before using the feature in the app