Inverted result patch
Addition of more comments.
Correction of an error on the way to get Frword, the structure is not alway the same.
Addition of inverted result, for small dictionnary when you search a word in La to Lb wordreference also show word in Lb that have a translation which contain word in La.
It's now handled.
Some more small fix.